a shot at using Box<Borrow<...>> in ExpressionInner
Because trait objects are invariant on their lifetimes, this requires us
to introduce a second lifetime param on Expression. See:
https://www.reddit.com/r/rust/comments/4eel2v/when_can_an_object_with_a_static_lifetime

Smart money says that invariance issue is going to break other cases,
maybe something like trying to pipe a shared expression with an owned
one. Anyway, the whole thing is starting to look a lot more complicated
than it needs to be. Hopefully I can rip a lot of this out.

The only ioarg types so far that really need borrowing are Files and
very large input Vecs. The answer could be to make everything else (the
Expressions themselves, in particular) stop trying to be so clever and
Just Make Copies.