use Into instead of AsRef
The impls for From/Into with respect to the different string types are
actually really extensive. They cover the cases we were using AsRef for,
and they avoid making an unnecessary copy when the argument is already
owned.

One downside is that there aren't impls for &Vec (as opposed to &[T]),
so callers will need to use &* in some cases where & would previously
have worked. I'm not sure a (tiny) performance improvement by itself
would be worth the (small) drop in ergonomics, but using a more
idiomatic trait is definitely worth it.