stop using error-chain, and always return io::Error instead
We're going to return shared references to output in the upcoming Handle
API, and that didn't mesh well with the Status error type. Also the
unchecked() method has replaced most of the use case for that error.
Getting rid of it is simpler, and the Utf8 error type also folds
naturally into io:Error.